requirements.txt
streamlit
polars
langchain
langchain-google-genai
langchain-experimental
plotly
python-dotenv
openpyxl